Created PostPage.js
Created this page after refactoring and moving all the parts in app.js into PostPage.js
